+++ apps/frontend/src/pages/auth/auth.tsx
@@ -1,9 +1,10 @@
 import RippleEffectBg from "~/assets/img/ripple-effect-bg.svg?react";
 import RippleEffectBg2 from "~/assets/img/ripple-effect-bg2.svg?react";
-import { Link } from "~/libs/components/components.js";
+import { Link, Navigate } from "~/libs/components/components.js";
 import { AppRoute } from "~/libs/enums/enums.js";
 import {
 	useAppDispatch,
+	useAppSelector,
 	useCallback,
 	useLocation,
 } from "~/libs/hooks/hooks.js";
@@ -19,6 +20,10 @@ import styles from "./styles.module.css";
 const Auth: React.FC = () => {
 	const dispatch = useAppDispatch();
 	const { pathname } = useLocation();
+
+	const { user } = useAppSelector(({ auth }) => ({
+		user: auth.user,
+	}));
 
 	const handleSignInSubmit = useCallback(
 		(payload: UserSignInRequestDto): void => {
@@ -48,10 +53,17 @@ const Auth: React.FC = () => {
 		return null;
 	};
 
+	const hasUser = Boolean(user);
+
+	if (hasUser) {
+		return <Navigate replace to={AppRoute.ROOT} />;
+	}
+
 	return (
 		<div className={styles["auth-container"]}>
 			<RippleEffectBg className={styles["ripple-effect__background1"]} />
 			<RippleEffectBg2 className={styles["ripple-effect__background2"]} />
+			<div className={styles["white-dots"]} />
 			<div className={styles["form-container"]}>
 				<div className={styles["form-header"]}>
 					<div className={styles["form-header__logo-container"]}>