+++ apps/mobile/src/libs/components/components.tsx
@@ -3,9 +3,12 @@ export { BackgroundWrapper } from "./background-wrapper/background-wrapper";
 export { Button } from "./button/button";
 export { Checkbox } from "./checkbox/checkbox";
 export { GradientText } from "./gradient-text/gradient-text";
+export { Icon } from "./icon/icon";
+export { IconButton } from "./icon-button/icon-button";
 export { Input } from "./input/input";
 export { Link } from "./link/link";
 export { LoaderWrapper } from "./loader-wrapper/loader-wrapper";
+export { Planet } from "./planet/planet";
 export { ScreenWrapper } from "./screen-wrapper/screen-wrapper";
 export { Text } from "./text/text";
 export { ToastMessage } from "./toast-message/toast-message";
@@ -16,6 +19,9 @@ export {
 	KeyboardAvoidingView,
 	Pressable,
 	ScrollView,
+	TextInput,
+	TouchableOpacity,
 	View,
 } from "react-native";
 export { LinearGradient } from "react-native-linear-gradient";
+export { default as MaterialIcon } from "react-native-vector-icons/MaterialIcons";